# codeintel: Reduce memory when calculating commit graph