… enterprise/internal/codeintel/stores/dbstore/dumps_test.go
… enterprise/internal/codeintel/stores/dbstore/helpers_test.go
… enterprise/internal/codeintel/stores/dbstore/references_test.go